Update DatabaseUtility.php
Added validation of $query and $prefix to prevent SQL injection attacks
Used mysql_real_escape_string to escape special characters in the checkSQL function

DESCRIPTION:
The original code did not have any measures in place to prevent SQL injection attacks. To address this, I added validation of the $query and $prefix parameters in the prefixQuery function using the checkSQL function. The checkSQL function uses mysql_real_escape_string to escape special characters in the input string, and then checks if the escaped string is different from the original string. If it is different, it means that special characters were found and escaped, and therefore the input string is not safe. 

Observation: 
I am participating in the HackerOne program and have also uploaded my findings there. I hope my contributions will be of assistance in improving security measures.